Initial commit.
This project is taken from the "Samples" in Urho3D source code.

However, in Urho3D source code it is not easy to understand
which files belong to this specific example and which files are
used by other sample projects.

This project contains only the minimal amount of files needed
to run the Urho2DPlatformer example, so that it can be compared
to other 2D platformers created with other engines.